+++ src/main/java/Dominio/WeatherAdapter.java
@@ -1,5 +1,5 @@
 package Dominio;
 
 public interface WeatherAdapter {
-	public int darTemperaturaActual(int day);
+	public float darTemperaturaActual(int day);
 }
